Implement full rounds in Game state
Now when we enter the game state, we start in the begin round state,
ensure that everything is reset, and then proceed into the game.

The game starts in the begin round state, which checks to see if the
game is over and either goes to the maze generation for this round or
counts this as game over.

At the end of the final ball drop we go to the end of round state,
where we update the current round and then jump to the begin round
state.

This also makes sure that the automatic arrows generate during maze
generation if this is the last round of a multi-round game.