Add computer ball toggle
Instead of having keys to selectively view either ball set, on the
player's turn they can press a key to view the computer balls instead,
to evaluate what the computer can possibly do.

This is handled as an extra state, so that when the computer balls are
visible, the player cannot move or play a ball.