Use lerping in the score
Finally all of the commits come together.

Now when a ball has stopped moving and is going to be going away,
instead of vanishing it, we use the new score start animation, which
is a partial animation that shows the ball starting to vanish.

Once that animation finishes (with the ball still half visible), we
trigger the callback, which scores the ball, telling it to lerp to the
location where the associated score is being displayed and also finish
the vanish.

The result is that the ball starts to vanish and then flies over to
the score, allowing a bit of a visual clue that you got points for
that ball.