Fix timing on ball lerping animations
Instead of having the code that vanishes the ball half vanish it with
a short animation, then lerp the ball while playing another animation,
we instead vanish directly but apply the lerp right away as well.

This way the score updates roughly at the same time as the ball
arrives there, which makes for a nicer visual look.

In the case of a blocked ball this is not entirely true, but the score
is only updating a few frames early and there are a lot of that going
on all at once, so it blends and fools the eye.