Display Billboard for no moves left
When either player has no moves left (which includes at the end of the
round), we display a billboard to say that.

Like the change for the start player announcement, this uses the next
state to store where to go next, and like the final ball drop we go
into another state to wait out the billboard timer.