Use constant for billboard delay time
This seems nicer than hard coding it; it's easier to either change all
values, or find places where specific billboards are waiting so that
we can change them.

Might be interesting to make a duration property on the Billboard
entity and then have it set as part of the show() method and queried.

Then we could say:
    this._billboard.show ("text", 35);
    if (this._state.hasElapsed (this._billboard.duration))
        this._billboard.hide ();

To track per billboard timings.